🚚MOVE: 게시글 댓글 도메인 분리하기(#31)
이제 게시글과 댓글 도메인을 분리하자. 그리고 잘못 분리된 파일을 다시 제대로 도메인에 따라서 정상적으로 변경하자. 그리고 댓글과 게시글을 분리한다.
