+++ src/main/java/com/dedication/force/article/repository/ArticleRepository.java
@@ -1,6 +1,6 @@
 package com.dedication.force.article.repository;
 
-import com.dedication.force.member.domain.entity.Article;
+import com.dedication.force.article.domain.entity.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArticleRepository extends JpaRepository<Article, Long> {
