+++ src/main/java/com/dedication/force/article/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package com.dedication.force.article.repository;
 
-import com.dedication.force.member.domain.entity.Category;
+import com.dedication.force.article.domain.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
