+++ src/main/java/com/dedication/force/comment/domain/entity/Comment.java
@@ -1,26 +1,22 @@
-package com.dedication.force.member.domain.entity;
+package com.dedication.force.comment.domain.entity;
 
-import com.dedication.force.article.domain.entity.Member;
+import com.dedication.force.member.domain.entity.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 import java.time.ZonedDateTime;
-import java.util.Set;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class Article {
+public class Comment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "article_id", nullable = false, updatable = false)
+    @Column(name = "comment_id", nullable = false, updatable = false)
     private Long id;
 
-    @Column(length = 50, nullable = false)
-    private String title;
-
-    @Column(length = 3000, nullable = false)
+    @Column(length = 200, nullable = false)
     private String content;
 
     @Column(nullable = false)
@@ -32,7 +28,4 @@ public class Article {
     @JoinColumn(name = "member_id")
     @ManyToOne(fetch = FetchType.LAZY)
     private Member member;
-
-    @OneToMany(mappedBy = "article", cascade = CascadeType.ALL, orphanRemoval = true)
-    private Set<ArticleCategory> articleCategories;
 }
