+++ src/main/java/com/dedication/force/comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
-package com.dedication.force.article.repository;
+package com.dedication.force.comment.repository;
 
-import com.dedication.force.member.domain.entity.Comment;
+import com.dedication.force.comment.domain.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
