+++ src/main/java/com/dedication/force/common/security/CustomUserDetails.java
@@ -1,6 +1,6 @@
 package com.dedication.force.common.security;
 
-import com.dedication.force.article.domain.entity.Member;
+import com.dedication.force.member.domain.entity.Member;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
