+++ src/main/java/com/dedication/force/common/security/CustomUserDetailsService.java
@@ -1,6 +1,6 @@
 package com.dedication.force.common.security;
 
-import com.dedication.force.article.domain.entity.Member;
+import com.dedication.force.member.domain.entity.Member;
 import com.dedication.force.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.InternalAuthenticationServiceException;
