+++ src/main/java/com/dedication/force/member/domain/dto/LoginMemberRequest.java
@@ -1,12 +1,10 @@
-package com.dedication.force.article.domain.dto;
+package com.dedication.force.member.domain.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
-import lombok.Builder;
 
-@Builder
-public record AddMemberRequest(
+public record LoginMemberRequest(
         @NotBlank(message = "이메일: 필수 정보입니다.")
         @Pattern(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,6}$", message = "이메일: 유효한 이메일 주소를 입력해주세요.")
         @Size(min = 6, max = 32, message = "이메일: 유효한 이메일 주소를 입력해주세요.")
@@ -16,12 +14,5 @@ public record AddMemberRequest(
         @Pattern(regexp = "(?=.*[0-9])(?=.*[a-zA-Z])(?=.*\\W)(?=\\S+$).{8,20}"
                 , message = "비밀번호: 8~20자 영문 대소문자, 숫자, 특수문자를 조합하여 작성해야 합니다.")
         @Size(min = 8, max = 20, message = "비밀번호: 8~20자 영문 대소문자, 숫자, 특수문자를 조합하여 작성해야 합니다.")
-        String password,
-
-        @NotBlank(message = "휴대전화번호: 필수 정보입니다.")
-        @Pattern(regexp = "^[0-9]{10,11}$", message = "휴대전화번호: 10~11자 숫자만 입력해주세요.")
-        @Size(min = 10, max = 11, message = "휴대전화번호: 10~11자 숫자만 입력해주세요.")
-        String phone
-) {
-
-}
+        String password
+) { }
