+++ src/main/java/com/dedication/force/member/domain/dto/LogoutRequest.java
@@ -1,3 +1,3 @@
-package com.dedication.force.article.domain.dto;
+package com.dedication.force.member.domain.dto;
 
 public record LogoutRequest(String refreshToken) { }
