+++ src/main/java/com/dedication/force/member/domain/entity/MemberRole.java
@@ -1,4 +1,4 @@
-package com.dedication.force.article.domain.entity;
+package com.dedication.force.member.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
