+++ src/main/java/com/dedication/force/member/repository/MemberRoleRepository.java
@@ -1,6 +1,6 @@
 package com.dedication.force.member.repository;
 
-import com.dedication.force.article.domain.entity.MemberRole;
+import com.dedication.force.member.domain.entity.MemberRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRoleRepository extends JpaRepository<MemberRole, Long> {
