+++ src/main/java/com/dedication/force/member/repository/RoleRepository.java
@@ -1,7 +1,7 @@
 package com.dedication.force.member.repository;
 
-import com.dedication.force.article.domain.entity.Role;
-import com.dedication.force.article.domain.entity.RoleType;
+import com.dedication.force.member.domain.entity.Role;
+import com.dedication.force.member.domain.entity.RoleType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
