+++ src/test/java/com/dedication/force/common/jwt/JwtTokenProviderTest.java
@@ -1,6 +1,6 @@
 package com.dedication.force.common.jwt;
 
-import com.dedication.force.article.domain.entity.RoleType;
+import com.dedication.force.member.domain.entity.RoleType;
 import com.dedication.force.common.jwt.dto.JwtTokenRequest;
 import io.jsonwebtoken.Claims;
 import org.assertj.core.api.Assertions;
